--- mongod
@@ -1,1 +1,0 @@
-mongod --bind_ip=$IP--dbpath=data --nojournal --rest "$@"
